python-server/requirements.txt
aiohttp==3.11.13
websockets==12.0
python-dotenv==1.0.1 